+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - (Fix) Corrected URL encoding for multiple resources
 - (Fix) `fmc_device`: add missing FTDv100 performance tier
+- (Fix) `fmc_vpn_s2s_endpoints`: import
 - (Enhancement) Add support for Security Cloud Control (SCC) Firewall Management Base URI
 
 ## 2.0.0-rc3